Reduce fieldset checkbox/radio duplication
These two classes act in a very-similar manner, much of the
functionality is duplicate. As all fieldset items are conditional, we
can safely remove the Traits::Conditional and just treat everything as
a FieldsetItem.

This is the same approach taken by the text fields and Traits::Input.

Refs #283
