+++ lib/govuk_design_system_formbuilder/traits/fieldset_item.rb
@@ -1,7 +1,37 @@
 module GOVUKDesignSystemFormBuilder
   module Traits
     module FieldsetItem
+      using PrefixableArray
+
+      def html
+        safe_join([item, @conditional])
+      end
+
     private
+
+      def class_prefix
+        %(#{brand}-#{input_type})
+      end
+
+      def item
+        tag.div(class: %(#{class_prefix}__item)) do
+          safe_join([input, label_element, hint_element])
+        end
+      end
+
+      def options
+        {
+          id: field_id(link_errors: @link_errors),
+          class: classes,
+          multiple: @multiple,
+          aria: { describedby: [hint_id] },
+          data: { 'aria-controls' => @conditional_id }
+        }
+      end
+
+      def classes
+        [%(#{class_prefix}__input)]
+      end
 
       def label_element
         @label_element ||= if @label.nil?
@@ -25,6 +55,27 @@ module GOVUKDesignSystemFormBuilder
 
       def hint_options
         { value: @value }.merge(fieldset_options)
+      end
+
+      def conditional_id
+        build_id('conditional')
+      end
+
+      def conditional_content(&block)
+        if (conditional_block_content = block_given? && block.call.presence)
+          @conditional    = conditional_container(conditional_block_content)
+          @conditional_id = conditional_id
+        end
+      end
+
+      def conditional_container(content)
+        tag.div(class: conditional_classes, id: conditional_id) do
+          capture { content }
+        end
+      end
+
+      def conditional_classes
+        %w(__conditional __conditional--hidden).prefix(class_prefix, delimiter: nil)
       end
     end
   end
